refactor(stations): split command and query services
Extract station command/query repositories, services, and service types so domain logic is easier to follow, while keeping thin compatibility facades during the migration.
