+++ apps/server/src/domain/stations/repository/station.repository.ts
@@ -1,18 +1,19 @@
 import { Effect, Layer } from "effect";
 
-import type {
-  PrismaClient,
-  Prisma as PrismaTypes,
-} from "generated/prisma/client";
+import type { PrismaClient, Prisma as PrismaTypes } from "generated/prisma/client";
 
 import { Prisma } from "@/infrastructure/prisma";
 
 import type { StationRepo } from "./station.repository.types";
 
-import { makeStationReadRepository } from "./read/station.read.repository";
-import { makeStationWriteRepository } from "./write/station.write.repository";
+import {
+  makeStationCommandRepository,
+} from "./station-command.repository";
+import {
+  makeStationQueryRepository,
+} from "./station-query.repository";
 
-export { toStationOrderBy } from "./station.repository.helpers";
+export { toStationOrderBy } from "./station-query.repository";
 export type { StationRepo } from "./station.repository.types";
 
 const makeStationRepositoryEffect = Effect.gen(function* () {
@@ -31,8 +32,8 @@ export function makeStationRepository(
   client: PrismaClient | PrismaTypes.TransactionClient,
 ): StationRepo {
   return {
-    ...makeStationReadRepository(client),
-    ...makeStationWriteRepository(client),
+    ...makeStationQueryRepository(client),
+    ...makeStationCommandRepository(client),
   };
 }
