+++ apps/server/src/http/shared/features/station.layers.ts
@@ -1,6 +1,5 @@
 import { Layer } from "effect";
 
-import { ReservationQueryRepositoryLive } from "@/domain/reservations";
 import {
   StationRepositoryLive,
   StationServiceLive,
@@ -13,17 +12,12 @@ export const StationReposLive = StationRepositoryLive.pipe(
   Layer.provide(PrismaLive),
 );
 
-export const StationReservationQueryReposLive = ReservationQueryRepositoryLive.pipe(
-  Layer.provide(PrismaLive),
-);
-
 export const StationServiceLayer = StationServiceLive.pipe(
-  Layer.provide(Layer.mergeAll(StationReposLive, StationReservationQueryReposLive, AgencyReposLive)),
+  Layer.provide(Layer.mergeAll(StationReposLive, AgencyReposLive)),
 );
 
 export const StationDepsLive = Layer.mergeAll(
   StationReposLive,
-  StationReservationQueryReposLive,
   StationServiceLayer,
   PrismaLive,
 );
